pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iocmng"
dynamic = ["version"]
description = "A pluggable task/job framework for IOC Manager applications with REST API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "INFN Beamline Controls" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "softioc",
    "cothread",
]

[project.optional-dependencies]
ophyd = [
    "ophyd",
    "infn_ophyd_hal",
]
kubernetes = [
    "kubernetes",
]
all = [
    "iocmng[ophyd,kubernetes]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "httpx",
    "black",
    "flake8",
    "build",
    "twine",
]

[project.scripts]
iocmng-server = "iocmng.api.app:run_server"

[project.urls]
Repository = "https://github.com/infn-epics/epik8s-beamline-controller"

[tool.setuptools.dynamic]
version = { attr = "iocmng.__version__" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
